% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{biol321Ids}
\alias{biol321Ids}
\title{Biol321 Course Ids}
\format{
A data frame with 30 rows and 1 variables:
\describe{
\item{\code{SIS.ID}}{character Course Ids}
}
}
\source{
\url{https://utexas.instructure.com/courses/1322979}
}
\usage{
biol321Ids
}
\description{
Current Biol321G Id's as of Fall 2021
}
\keyword{datasets}
