% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idToNumber.R
\name{idToNumber}
\alias{idToNumber}
\title{Convert strings to numbers for use in assigning random inputs based on unique identifiers}
\usage{
idToNumber(
  id,
  matchTable = c(letters, LETTERS, 0:9),
  formatWidth = nchar(length(matchTable))
)
}
\arguments{
\item{id}{a character vector containing strings to convert to numbers}

\item{matchTable}{a character vector with all characters present in id}

\item{formatWidth}{a length 1 integer vector used to justify idMatch.}
}
\value{
a named numeric vector of the input characters converted to the concatenated matching position within matchTable
}
\description{
Convert strings to numbers for use in assigning random inputs based on unique identifiers
}
\examples{
idToNumber("nbb624")
}
