% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rGene.R
\name{rGene}
\alias{rGene}
\title{Random gene selector}
\usage{
rGene(id, seed = 20211017, randomRow = F)
}
\arguments{
\item{id}{A length 1 character vector of your ID (should match a value in biol321Ids)}

\item{seed}{A length 1 numeric vector used to set pseudo-random assignments. This seed set per semester; do not change unless instructed to do so.}

\item{randomRow}{A length one logical vector. Only change this value if your Id is not in the biol321Ids object for some reason.}
}
\value{
The random humanGene row that was selected based on your ID
}
\description{
Random gene selector
}
\examples{
# Typical use
rGene("nbb624")
}
